baselines/naist-simulst/requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
audioread==3.0.0
-e git+ssh://git@github.com/facebookresearch/WavAugment.git@54afcdb00ccc852c2f030f239f8532c9562b550e#egg=augment
bitarray==2.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
contourpy==1.0.7
coverage==7.2.3
cycler==0.11.0
Cython==0.29.34
decorator==5.1.1
docker-pycreds==0.4.0
exceptiongroup==1.1.1
-e fairseq/.
ffmpeg-python==0.2.0
filelock==3.11.0
flake8==6.0.0
fonttools==4.39.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.53.0
gspread==5.8.0
huggingface-hub==0.13.4
hydra-core==1.0.7
idna==3.4
iniconfig==2.0.0
ipadic==1.0.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
lit==16.0.1
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mccabe==0.7.0
mecab-python3==1.0.5
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.0.6
openai-whisper==20230314
packaging==23.1
pandas==2.0.0
pathtools==0.1.2
Pillow==9.5.0
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
protobuf==4.22.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.3.1
pytest-cov==4.0.0
pytest-flake8==1.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.98
sentry-sdk==1.19.1
setproctitle==1.3.2
-e SimulEval/.
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
TextGrid==1.5
threadpoolctl==3.1.0
tiktoken==0.3.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
tornado==6.2
tqdm==4.64.1
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.14.2
Werkzeug==2.2.3
